Remove function from Utils
